% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_grid}
\alias{create_grid}
\title{Plot a grid of the US with maximum resolution specified.
Coordinate grid points are in Mercator projection.}
\usage{
create_grid(resolution_X = 50, resolution_Y = 50)
}
\arguments{
\item{resolution_X}{an integer of the maximum resolution with respect to longitude to be plotted.}

\item{resolution_Y}{an integer of the maximum resolution with respect to latitude to be plotted.}
}
\value{
a set of points at the specified resolution with a logical column representing whether the point falls within the contiguous USA.
}
\description{
Plot a grid of the US with maximum resolution specified.
Coordinate grid points are in Mercator projection.
}
\examples{
# get a plot of the USA at 30 maximum points with respect to longitude and 20 with respect to latitude
point_map <- create_grid(resolution_X = 30, resolution_Y=20)
plot(point_map)
}
