% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-functions.R
\name{get_mesonet_weather}
\alias{get_mesonet_weather}
\title{Get Mesonet Weather Data}
\usage{
get_mesonet_weather(station, year1, month1, day1, year2, month2, day2, tz)
}
\description{
This function gets weather data from MESONET for a particular
weather station (find this using \link{get_station_info}), then converts
the data into daily averages in Celcius, knots, and 0.1xmm precipitation.
This function requires a lot of arguments, and it really exists just to
test that you have the values specified correctly before you use the next
function, \link{get_mesonet_weather_df}, on a dataframe with columns
that are named correctly.
}
