% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-functions.R
\name{get_station_info}
\alias{get_station_info}
\title{Get Mesonet Station Info}
\usage{
get_station_info(df, decdegrees = 0.1, loc_key = "Location_code")
}
\arguments{
\item{df}{A dataframe or tbl_df object that contains columns 'Latitude'
and 'Longitude', and some grouping column to specify distinct locations.}

\item{decdegrees}{The distance, in decimal degrees,}

\item{loc_key}{The name of the grouping column that specifies distinct
locations in the dataframe.}
}
\value{
a dataframe containing details for all Mesonet stations that are
within some number of decimal degrees of the latitudes and longitudes
in your input dataset.
}
\description{
The first step for getting weather data for a location is to
find which weather stations are closest to that location. This function
will return a dataframe of Mesonet weather stations that are within some
number of decimal degrees of some dataframe containing latitudes and
longitudes that you provide.
}
\examples{
    \dontrun{get_station_info(df = bean_data, loc_key = "Location_code")}
    \dontrun{get_station_info(df = bean_data, decdegrees = 0.5,
    loc_key = "Location_code")}

}
